Issue #71 - Implement spectrum projections
Initial setup
