+++ src/main.rs
@@ -14,8 +14,8 @@ use clap::Parser;
 use portman_client;
 use rest::{
     apply, channel, data_processing, evbunpack, exit, filter, fit, fold, gates, getstats,
-    integrate, mirror_list, rest_parameter, ringversion, sbind, shm, spectrum, spectrumio, traces,
-    unbind, unimplemented, version,
+    integrate, mirror_list, project, rest_parameter, ringversion, sbind, shm, spectrum, spectrumio,
+    traces, unbind, unimplemented, version,
 };
 use sharedmem::{binder, mirror};
 use std::env;
@@ -239,7 +239,7 @@ fn rocket() -> _ {
                 unimplemented::pman_clone
             ],
         )
-        .mount("/spectcl/project", routes![unimplemented::project])
+        .mount("/spectcl/project", routes![project::project])
         .mount(
             "/spectcl/pseudo",
             routes![
